+++ src/graphql/queries.ts
@@ -90,7 +90,7 @@ export type NetworkCostRow = {
   network_level: Scalars["String"]["output"];
   operator_id: Scalars["Int"]["output"];
   operator_name: Scalars["String"]["output"];
-  rate: Scalars["Float"]["output"];
+  rate?: Maybe<Scalars["Float"]["output"]>;
   year: Scalars["Int"]["output"];
 };
 
@@ -137,9 +137,9 @@ export enum ObservationKind {
 
 export type OperationalStandardsCompliance = {
   __typename: "OperationalStandardsCompliance";
-  francsRule: Scalars["String"]["output"];
+  francsRule?: Maybe<Scalars["String"]["output"]>;
   operatorsFrancsPerInvoice: Array<OperationalStandardsOperatorFrancs>;
-  timelyPaperSubmission: Scalars["Boolean"]["output"];
+  timelyPaperSubmission?: Maybe<Scalars["Boolean"]["output"]>;
 };
 
 export type OperationalStandardsData = {
@@ -163,22 +163,22 @@ export type OperationalStandardsOperator = {
 
 export type OperationalStandardsOperatorFrancs = {
   __typename: "OperationalStandardsOperatorFrancs";
-  francsPerInvoice: Scalars["Float"]["output"];
+  francsPerInvoice?: Maybe<Scalars["Float"]["output"]>;
   operatorId: Scalars["String"]["output"];
   year: Scalars["String"]["output"];
 };
 
 export type OperationalStandardsOperatorNotification = {
   __typename: "OperationalStandardsOperatorNotification";
-  days: Scalars["Int"]["output"];
+  days?: Maybe<Scalars["Int"]["output"]>;
   operatorId: Scalars["String"]["output"];
   year: Scalars["String"]["output"];
 };
 
 export type OperationalStandardsServiceQuality = {
   __typename: "OperationalStandardsServiceQuality";
-  informingCustomersOfOutage: Scalars["Boolean"]["output"];
-  notificationPeriodDays: Scalars["Int"]["output"];
+  informingCustomersOfOutage?: Maybe<Scalars["Boolean"]["output"]>;
+  notificationPeriodDays?: Maybe<Scalars["Int"]["output"]>;
   operatorsNotificationPeriodDays: Array<OperationalStandardsOperatorNotification>;
 };
 
@@ -458,8 +458,8 @@ export type StabilityDataRow = {
   __typename: "StabilityDataRow";
   operator_id: Scalars["Int"]["output"];
   operator_name: Scalars["String"]["output"];
-  total: Scalars["Float"]["output"];
-  unplanned: Scalars["Float"]["output"];
+  total?: Maybe<Scalars["Float"]["output"]>;
+  unplanned?: Maybe<Scalars["Float"]["output"]>;
   year: Scalars["Int"]["output"];
 };
 
@@ -552,7 +552,7 @@ export type TariffRow = {
   operator_id: Scalars["Int"]["output"];
   operator_name: Scalars["String"]["output"];
   period: Scalars["Int"]["output"];
-  rate: Scalars["Float"]["output"];
+  rate?: Maybe<Scalars["Float"]["output"]>;
 };
 
 export type TariffsData = {
@@ -1021,23 +1021,23 @@ export type OperationalStandardsQuery = {
     };
     serviceQuality: {
       __typename: "OperationalStandardsServiceQuality";
-      notificationPeriodDays: number;
-      informingCustomersOfOutage: boolean;
+      notificationPeriodDays?: number | null;
+      informingCustomersOfOutage?: boolean | null;
       operatorsNotificationPeriodDays: Array<{
         __typename: "OperationalStandardsOperatorNotification";
         operatorId: string;
-        days: number;
+        days?: number | null;
         year: string;
       }>;
     };
     compliance: {
       __typename: "OperationalStandardsCompliance";
-      francsRule: string;
-      timelyPaperSubmission: boolean;
+      francsRule?: string | null;
+      timelyPaperSubmission?: boolean | null;
       operatorsFrancsPerInvoice: Array<{
         __typename: "OperationalStandardsOperatorFrancs";
         operatorId: string;
-        francsPerInvoice: number;
+        francsPerInvoice?: number | null;
         year: string;
       }>;
     };
@@ -1060,7 +1060,7 @@ export type NetworkCostsQuery = {
     yearlyData: Array<{
       __typename: "NetworkCostRow";
       year: number;
-      rate: number;
+      rate?: number | null;
       operator_id: number;
       operator_name: string;
       network_level: string;
@@ -1084,7 +1084,7 @@ export type EnergyTariffsQuery = {
     yearlyData: Array<{
       __typename: "TariffRow";
       period: number;
-      rate: number;
+      rate?: number | null;
       operator_id: number;
       operator_name: string;
       category: any;
@@ -1108,7 +1108,7 @@ export type NetTariffsQuery = {
     yearlyData: Array<{
       __typename: "TariffRow";
       period: number;
-      rate: number;
+      rate?: number | null;
       operator_id: number;
       operator_name: string;
       category: any;
@@ -1135,8 +1135,8 @@ export type SaidiQuery = {
     yearlyData: Array<{
       __typename: "StabilityDataRow";
       year: number;
-      total: number;
-      unplanned: number;
+      total?: number | null;
+      unplanned?: number | null;
       operator_id: number;
       operator_name: string;
     }>;
@@ -1162,8 +1162,8 @@ export type SaifiQuery = {
     yearlyData: Array<{
       __typename: "StabilityDataRow";
       year: number;
-      total: number;
-      unplanned: number;
+      total?: number | null;
+      unplanned?: number | null;
       operator_id: number;
       operator_name: string;
     }>;